refactor: move resource description into each resource method
