+++ bundle/config/resources/description.go
@@ -1,0 +1,13 @@
+package resources
+
+type ResourceDescription struct {
+	// Singular and plural name when used to refer to the configuration.
+	SingularName string
+	PluralName   string
+
+	// Singular and plural title when used in summaries / terminal UI.
+	SingularTitle string
+	PluralTitle   string
+
+	TerraformResourceName string
+}
