+++ bundle/config/resources/mlflow_experiment.go
@@ -38,8 +38,14 @@ func (s *MlflowExperiment) Exists(ctx context.Context, w *databricks.WorkspaceCl
 	return true, nil
 }
 
-func (j *MlflowExperiment) ResourceType() string {
-	return "experiment"
+func (j *MlflowExperiment) ResourceDescription() ResourceDescription {
+	return ResourceDescription{
+		SingularName:          "experiment",
+		PluralName:            "experiments",
+		SingularTitle:         "Experiment",
+		PluralTitle:           "Experiments",
+		TerraformResourceName: "databricks_mlflow_experiment",
+	}
 }
 
 func (s *MlflowExperiment) TerraformResourceName() string {
