+++ bundle/config/resources/pipeline.go
@@ -38,8 +38,14 @@ func (p *Pipeline) Exists(ctx context.Context, w *databricks.WorkspaceClient, id
 	return true, nil
 }
 
-func (j *Pipeline) ResourceType() string {
-	return "pipeline"
+func (j *Pipeline) ResourceDescription() ResourceDescription {
+	return ResourceDescription{
+		SingularName:          "pipeline",
+		PluralName:            "pipelines",
+		SingularTitle:         "Pipeline",
+		PluralTitle:           "Pipelines",
+		TerraformResourceName: "databricks_pipeline",
+	}
 }
 
 func (p *Pipeline) TerraformResourceName() string {
