+++ bundle/config/resources/quality_monitor.go
@@ -42,8 +42,14 @@ func (s *QualityMonitor) Exists(ctx context.Context, w *databricks.WorkspaceClie
 	return true, nil
 }
 
-func (*QualityMonitor) ResourceType() string {
-	return "quality_monitor"
+func (*QualityMonitor) ResourceDescription() ResourceDescription {
+	return ResourceDescription{
+		SingularName:          "quality_monitor",
+		PluralName:            "quality_monitors",
+		SingularTitle:         "Quality Monitor",
+		PluralTitle:           "Quality Monitors",
+		TerraformResourceName: "databricks_quality_monitor",
+	}
 }
 
 func (s *QualityMonitor) TerraformResourceName() string {
